+++ pennylane/decompositions/decomposition.py
@@ -23,7 +23,7 @@ from rustworkx.visit import DijkstraVisitor, StopSearch, PruneSearch
 
 from pennylane.operation import Operator
 
-from .compressed_op import CompressedResourceOp
+from .. import CompressedResourceOp
 from .decomposition_rule import DecompositionRule
 from .resources import Resources
